launch: Change linux/snp module to linux/ module
